pi_code/requirements.txt
pygame==2.5.2
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
RPi.GPIO==0.7.1
